Revert "[CORE07684] 2.4.y.z/uppy integration trunk"
