+++ projects/v3/src/app/components/uppy-uploader/uppy-uploader.service.ts
@@ -55,22 +55,4 @@ export class UppyUploaderService {
 
     return modal;
   }
-
-  // extract response from tus upload XHR state
-  extractResponseData(response: {
-    _xhr: {
-      response: string;
-    };
-  }): {
-    path: string;
-    bucket: string;
-  } {
-    const res = response?._xhr?.response;
-    const data = JSON.parse(res);
-
-    // eslint-disable-next-line no-console
-    console.log('uppy-xhr', data);
-
-    return data;
-  }
 }
