+++ projects/v3/src/app/pages/settings/settings.page.html
@@ -15,7 +15,7 @@
     <ng-container *ngIf="!isMobile; else mobileProfile">
       <div class="avatar-and-profile">
         <ion-avatar class="avatar" slot="start">
-          <img *ngIf="!imageUpdating" [src]="profile?.avatar" alt="user profile" i18n-alt [appFallbackImage]="'assets/loading.gif'">
+          <img *ngIf="!imageUpdating" [src]="profile?.image" alt="user profile" i18n-alt [appFallbackImage]="'assets/loading.gif'">
           <p *ngIf="imageUpdating" class="loading-icon">
             <ion-spinner name="bubbles"></ion-spinner>
           </p>
@@ -30,7 +30,7 @@
     <ng-template #mobileProfile>
       <ion-item>
         <ion-avatar class="avatar" slot="start">
-          <img *ngIf="!imageUpdating" [src]="profile?.avatar || '/assets/loading.gif'" alt="user profile">
+          <img *ngIf="!imageUpdating" [src]="profile?.image || '/assets/loading.gif'" alt="user profile">
           <p *ngIf="imageUpdating" class="loading-icon">
             <ion-spinner name="bubbles"></ion-spinner>
           </p>
