--- projects/v3/src/app/services/types.ts
@@ -1,4 +1,0 @@
-export interface Response {
-  success: boolean;
-  message: string;
-}
